part 3 added

+++ src/app/modules/student/student.controller.ts
@@ -2,6 +2,8 @@ import { studentServices } from './student.service';
 import sendResponse from '../../utils/sendResponse';
 import httpStatus from 'http-status';
 import catchAsync from '../../utils/catchAsync';
+import { Student } from './student.model';
+import AppError from '../../errors/appError';
 
 const getAllStudents = catchAsync(async (req, res) => {
   const result = await studentServices.getAllStudentsFromDB();
@@ -16,6 +18,11 @@ const getAllStudents = catchAsync(async (req, res) => {
 
 const getStudentById = catchAsync(async (req, res) => {
   const { studentId } = req.params;
+  const isUserExist = await Student.isUserExists(studentId);
+  if (!isUserExist) {
+    throw new AppError(httpStatus.BAD_REQUEST, 'Student not exist!');
+  }
+
   const result = await studentServices.getStudentByIdFromDB(studentId);
 
   sendResponse(res, {
@@ -26,8 +33,30 @@ const getStudentById = catchAsync(async (req, res) => {
   });
 });
 
+const updateStudentIntoDB = catchAsync(async (req, res) => {
+  const { studentId } = req.params;
+  const { student } = req.body;
+  const isUserExist = await Student.isUserExists(studentId);
+  if (!isUserExist) {
+    throw new AppError(httpStatus.BAD_REQUEST, 'Student not exist!');
+  }
+
+  const result = await studentServices.updateStudentIntoDB(studentId, student);
+
+  sendResponse(res, {
+    statusCode: httpStatus.OK,
+    success: true,
+    message: 'Student is updated successfully!',
+    data: result,
+  });
+});
+
 const deleteStudent = catchAsync(async (req, res) => {
   const { studentId } = req.params;
+  const isUserExist = await Student.isUserExists(studentId);
+  if (!isUserExist) {
+    throw new AppError(httpStatus.BAD_REQUEST, 'Student not exist!');
+  }
   const result = await studentServices.deleteStudentFromDB(studentId);
 
   sendResponse(res, {
@@ -42,4 +71,5 @@ export const studentControllers = {
   getAllStudents,
   getStudentById,
   deleteStudent,
+  updateStudentIntoDB,
 };
